refactor: move hub and client implementation to chat package
